Fix many compiler warnings for gcc -Wextra for ogr, fmts and alg (https://trac.osgeo.org/gdal/ticket/5414)
git-svn-id: https://svn.osgeo.org/gdal/branches/1.11/gdal@27729 f0d54148-0727-0410-94bb-9a71ac55c965

+++ frmts/ceos/ceosopen.c
@@ -275,7 +275,7 @@ CEOSImage * CEOSOpen( const char * pszFilename, const char * pszAccess )
     if( psImage->nImageRecLength <= 0 ||
         psImage->nPrefixBytes < 0 ||
         psImage->nBands > INT_MAX / psImage->nImageRecLength ||
-        psImage->nBands > INT_MAX / sizeof(int))
+        (size_t)psImage->nBands > INT_MAX / sizeof(int))
     {
         CEOSDestroyRecord( psRecord );
         CEOSClose( psImage );
